requirements.txt
streamlit==1.30.0
gspread==5.12.0
google-auth==2.27.0
